Add files via upload
For my final year project, i had built a MIDI powered DJ Controller. I created this code that allows for 4 potentiometers, 3 sliders and 10 buttons that work with any DJ software. However, after purchasing a vast amount of Arduino kit, i realised that the Arduino Mega I was using wasn't MIDI compatible. To overcome this, i used a virtual Midi port (Midi Pipe/LoopMidi), that allowed the information to be sent through a Midi channel and direct to the DJ software that was being used.
The only issue i encountered at the final stage was a very slight delay, due to the information being transferred through the virtual port.
Some very minor adjustments could be made to the code to make the fader and potentiometer response a lot tighter.
However the final result was great as i received a 2:1 grade for the Project